requirements.txt
Flask==2.2.2
pycryptodome==3.17
Requests==2.31.0
pyinstaller==6.2.0
pygame==2.5.2
cryptography==41.0.7
jwt==1.3.1
Flask-SQLAlchemy==3.0.0